+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* A very basic calendar widget to show month view or pick any spacific date.
+
+
+## 0.0.7
+
+* Option to hide the current displayed day.